add env.yaml and decoupled version

+++ img2dataset/main.py
@@ -13,7 +13,7 @@ from .writer import (
     DummySampleWriter,
 )
 from .reader import Reader
-from .downloader import Downloader
+from .downloader import Downloader, DummyDownloader
 from .distributor import multiprocessing_distributor, pyspark_distributor
 import fsspec
 import sys
@@ -54,7 +54,12 @@ def download(
     np_path = None,
     index_path=None,
     enable_faiss_memory_mapping=False,
-    k=20
+    k=20,
+    only_reader=False,
+    start_input_file=0,
+    max_input_file=None,
+    benchmark=False,
+    only_ids=False
 ):
     """Download is the main entry point of img2dataset, it uses multiple processes and download multiple files"""
     config_parameters = dict(locals())
@@ -75,31 +80,35 @@ def download(
     if not fs.exists(tmp_dir):
         fs.mkdir(tmp_dir)
 
-    def signal_handler(signal_arg, frame):  # pylint: disable=unused-argument
-        try:
-            fs.rm(tmp_dir, recursive=True)
-        except Exception as _:  # pylint: disable=broad-except
-            pass
-        logger_process.terminate()
-        sys.exit(0)
+    if not only_reader:
+        print('Setting up signal handler to clean up tmp_dir')
+        def signal_handler(signal_arg, frame):  # pylint: disable=unused-argument
+            try:
+                fs.rm(tmp_dir, recursive=True)
+            except Exception as _:  # pylint: disable=broad-except
+                pass
+            logger_process.terminate()
+            sys.exit(0)
 
-    signal.signal(signal.SIGINT, signal_handler)
+        signal.signal(signal.SIGINT, signal_handler)
 
     save_caption = caption_col is not None
 
     fs, output_path = fsspec.core.url_to_fs(output_folder)
 
-    if not fs.exists(output_path):
-        fs.mkdir(output_path)
-        start_shard_id = 0
-    else:
-        existing_top_level_files = [x for x in fs.glob(output_path + "/*") if x != tmp_dir and "stats" not in x]
-        if len(existing_top_level_files) == 0:
-            start_shard_id = 0
-        else:
-            start_shard_id = (
-                max([int(x.split("/")[-1].split(".")[0]) for x in existing_top_level_files if x != tmp_dir]) + 1
-            )
+    start_shard_id = 0
+
+    # if not fs.exists(output_path):
+    #     fs.mkdir(output_path)
+    #     start_shard_id = 0
+    # else:
+    #     existing_top_level_files = [x for x in fs.glob(output_path + "/*") if x != tmp_dir and "stats" not in x]
+    #     if len(existing_top_level_files) == 0:
+    #         start_shard_id = 0
+    #     else:
+    #         start_shard_id = (
+    #                 max([int(x.split("/")[-1].split(".")[0]) for x in existing_top_level_files if x != tmp_dir]) + 1
+    #         )
 
     logger_process.start_shard_id = start_shard_id
     logger_process.start()
@@ -116,7 +125,11 @@ def download(
         np_path=np_path,
         index=index_path,
         enable_faiss_memory_mapping=enable_faiss_memory_mapping,
-        k=k
+        k=k,
+        start_input_file=start_input_file,
+        max_input_file=max_input_file,
+        benchmark=benchmark,
+        only_ids=only_ids
     )
 
     if output_format == "webdataset":
@@ -141,21 +154,26 @@ def download(
         disable_all_reencoding=disable_all_reencoding,
     )
 
-    downloader = Downloader(
-        sample_writer_class=sample_writer_class,
-        resizer=resizer,
-        thread_count=thread_count,
-        save_caption=save_caption,
-        extract_exif=extract_exif,
-        output_folder=output_folder,
-        column_list=reader.column_list,
-        timeout=timeout,
-        number_sample_per_shard=number_sample_per_shard,
-        oom_shard_count=oom_shard_count,
-        compute_md5=compute_md5,
-        retries=retries,
-        load_np=np_path is not None
-    )
+    if only_reader:
+        downloader = DummyDownloader()
+    else:
+        downloader = Downloader(
+            sample_writer_class=sample_writer_class,
+            resizer=resizer,
+            thread_count=thread_count,
+            save_caption=save_caption,
+            extract_exif=extract_exif,
+            output_folder=output_folder,
+            column_list=reader.column_list,
+            timeout=timeout,
+            number_sample_per_shard=number_sample_per_shard,
+            oom_shard_count=oom_shard_count,
+            compute_md5=compute_md5,
+            retries=retries,
+            load_np=np_path is not None,
+            only_ids=reader.only_ids
+        )
+
 
     print("Starting the downloading of this file")
     if distributor == "multiprocessing":
@@ -172,7 +190,8 @@ def download(
         subjob_size,
     )
     logger_process.join()
-    fs.rm(tmp_dir, recursive=True)
+    if not only_reader:
+        fs.rm(tmp_dir, recursive=True)
 
 
 def main():
